+++ project/player.php
@@ -1,28 +1,18 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="utf-8">
-    <link   href="css/bootstrap.min.css" rel="stylesheet">
-    <script src="js/bootstrap.min.js"></script>
+	<?php
+		include 'database.php';
+		Database::drawHeader(2);
+	?>
 </head>
  
 <body>
-	<div class="container">
-		<div class="btn-group btn-group-justified">
-				<a class="btn btn-primary" role="button" href="host.php" >Host</a>
-				<a class="btn btn-success" role="button" href="game.php">Game</a>
-				<a class="btn btn-warning" role="button" href="player.php" disabled="true" >Player</a>
-		</div>
-	</div>
     <div class="container">
             <div class="row">
                 <h3>Players</h3>
             </div>
             <div class="row">
-                <p>
-                    <a href="createplayer.php" class="btn btn-success">Create</a>
-                </p>
-                 
                 <table class="table table-striped table-bordered">
                       <thead>
                         <tr>
@@ -33,12 +23,10 @@
                         </tr>
                       </thead>
                       <tbody>
-                      </tbo<tbody>
-                      <?php
-                       include 'database.php';
-                       $pdo = Database::connect();
-                       $sql = 'SELECT * FROM players ORDER BY id DESC';
-                       foreach ($pdo->query($sql) as $row) {
+						<?php
+						$pdo = Database::connect();
+						$sql = 'SELECT * FROM players ORDER BY id DESC';
+						foreach ($pdo->query($sql) as $row) {
                                 echo '<tr>';
                                 echo '<td>'. $row['name'] . '</td>';
                                 echo '<td>'. $row['email_address'] . '</td>';
@@ -46,14 +34,19 @@
                                 echo '<td width=250>';
                                 echo '<a class="btn" href="readplayer.php?id='.$row['id'].'">Read</a>';
                                 echo ' ';
-                                echo '<a class="btn btn-success" href="updateplayer.php?id='.$row['id'].'">Update</a>';
-                                echo ' ';
-                                echo '<a class="btn btn-danger" href="deleteplayer.php?id='.$row['id'].'">Delete</a>';
-                                echo '</td>';
+								if($_SESSION['type'] == 0 | $_SESSION['type'] == 2){
+									if($_SESSION['user_id'] == $row['user_id'] | $_SESSION['type'] == 2){
+									echo '<a class="btn btn-success" href="updateplayer.php?id='.$row['id'].'">Update</a>';
+									echo ' ';
+									echo '<a class="btn btn-danger" href="deleteplayer.php?id='.$row['id'].'">Delete</a>';
+									}	
+								}
+								echo '</td>';
                                 echo '</tr>';
-                       }
-                       Database::disconnect();
-                      ?>
+						}
+						Database::disconnect();
+						?>
+					</tbody>
                 </table>
         </div>
     </div> <!-- /container -->
